+++ app/src/embeddings/__init__.py
@@ -1,11 +1,11 @@
 from src.embeddings.model import EmbeddingModel
-from src.embeddings.mpnet import MPNetEmbedding
+from app.src.embeddings.sentence_transformer import SentenceTransformerEmbedding
 from src.embeddings.openai import OpenAIEmbedding
 from src.embeddings.mock import MockEmbedding
 
 __all__ = [
     "EmbeddingModel",
-    "MPNetEmbedding",
+    "SentenceTransformerEmbedding",
     "OpenAIEmbedding",
     "MockEmbedding",
 ]
